test(onboard): split messaging coverage
